Allow pluggable sub-expression hashing in `IR::shallow_hash`
